requirements.txt
pygame==2.1.2
python-vlc==3.0.16120
supermemo2==2.0.0
PyYAML==6.0
python-dateutil==2.8.2